requirements.txt
requests==2.24.0
python-dotenv==0.14.0
redis~=3.5.3
pydash~=4.8.0
pymongo~=3.11.3
uvicorn~=0.11.8
pydantic~=1.6.1
starlette~=0.13.6
fastapi~=0.63.0
celery~=5.0.5
pendulum~=2.1.2